Fix libgomp tests.
2015-07-15  Maxim Blumenthal  <maxim.blumenthal@intel.com>

libgomp/
	* testsuite/libgomp.fortran/examples-4/simd-8.f90: (main): Change type
	of EPS parameter from integer to real.
	* testsuite/libgomp.fortran/examples-4/task_dep-5.f90: (check): Change
	type of EPS parameter from integer to real.

From-SVN: r225827

+++ libgomp/testsuite/libgomp.fortran/examples-4/task_dep-5.f90
@@ -44,7 +44,7 @@ contains
   subroutine check (N, A, B)
     integer :: N
     integer :: i, j
-    integer, parameter :: EPS = 0.000001
+    real, parameter :: EPS = 0.000001
     real, dimension(N,N) :: A, B
     real :: diff
     do i = 1, N
